重构了Finger了MCTS、MCTSNode、aiplayer、ais、fame、main、player、settings、source和模块，增加了核心逻辑、AI和UI的实现，重写了游戏逻辑方式。更新了README文档，增加了接口说明和。
