+++ src/AddOns/ai.py
@@ -1,0 +1,3 @@
+class AI:
+    def get_action(self, state):
+        return "all"
